LayoutTests/storage/indexeddb/objectStore-required-arguments-expected.txt

Test IndexedDB object store required arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;
PASS IDBDatabaseException == null is false
IDBCursor = window.IDBCursor || window.webkitIDBCursor;
PASS IDBCursor == null is false
IDBKeyRange = window.IDBKeyRange || window.webkitIDBKeyRange;
PASS IDBKeyRange == null is false
indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
objectStore = db.createObjectStore('foo');
PASS objectStore.put(); threw exception TypeError: Not enough arguments.
PASS objectStore.add(); threw exception TypeError: Not enough arguments.
PASS objectStore.delete(); threw exception TypeError: Not enough arguments.
PASS objectStore.get(); threw exception TypeError: Not enough arguments.
PASS objectStore.createIndex(); threw exception TypeError: Not enough arguments.
PASS objectStore.createIndex('foo'); threw exception TypeError: Not enough arguments.
PASS objectStore.index(); threw exception TypeError: Not enough arguments.
PASS objectStore.deleteIndex(); threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
